app: refactor icons
