+++ packages/app/src/components/icon/images.d.ts
