+++ packages/app/src/components/icon/images/autism-active.png
