+++ packages/app/src/components/icon/images/bipolar-active.png
