+++ packages/app/src/components/icon/images/child-active.png
