+++ packages/app/src/components/icon/images/child.png
