+++ packages/app/src/components/icon/images/depression-active.png
