+++ packages/app/src/components/icon/images/education-active.png
