+++ packages/app/src/components/icon/images/empathy-active.png
