+++ packages/app/src/components/icon/images/grief-active.png
