+++ packages/app/src/components/icon/images/home-active.png
