+++ packages/app/src/components/icon/images/home.png
