+++ packages/app/src/components/icon/images/instagram.png
