+++ packages/app/src/components/icon/images/kid-active.png
