+++ packages/app/src/components/icon/images/kid.png
