+++ packages/app/src/components/icon/images/logo.png
