+++ packages/app/src/components/icon/images/obsession-active.png
