+++ packages/app/src/components/icon/images/parent-active.png
