+++ packages/app/src/components/icon/images/parent.png
