+++ packages/app/src/components/icon/images/stress-active.png
