+++ packages/app/src/components/icon/images/teen-active.png
