+++ packages/app/src/components/icon/images/teen.png
