+++ packages/app/src/components/icon/images/telegram.png
